// src/modules/spot/dto/spot-order-info-response.dto.ts
export class SpotOrderInfoResponseDto {
  symbol: string;
  orderId: string;
  price: string;
  size: string;
  orderType: string;
  side: string;
  status: string;
  priceAvg: string;
  baseVolume: string;
  quoteVolume: string;
  feeCurrency: string;
  fee: string;
  cTime: string;
  uTime: string;
}